Be event

See merge request s05-webmobile1-sub3/S05P13C103!59

+++ backend/src/main/java/com/ssafy/api/request/LiveTitlePatchReq.java
@@ -15,7 +15,7 @@ import javax.persistence.Column;
 @ApiModel("LiveUpdate_titlePatchRequest")
 public class LiveTitlePatchReq {
 	@ApiModelProperty(name="방송 PK")
-	Long livepk;
+	Integer livepk;
 	@ApiModelProperty(name="방송 title")
 	String livetitle;
 }
