+++ backend/src/main/java/com/ssafy/db/entity/Product.java
@@ -20,16 +20,16 @@ public class Product{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "product_pk")
-    private Long id;
+    private Integer id;
 
     @Column(name="user_id_seller")
     private String userId;
 
     @Column(name = "live_id")
-    private Long liveId;
+    private Integer liveId;
 
     @Column(name = "category_pk")
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(name = "product_title")
     private String title;
@@ -51,7 +51,7 @@ public class Product{
     private Integer isSold;
 
     @Column(name = "product_is_live")
-    private Long isLive;
+    private Integer isLive;
 
     @JsonIgnore
     @OneToMany(
@@ -62,7 +62,7 @@ public class Product{
     private List<Image> image = new ArrayList<>();
 
     @Builder
-    public Product(String userId, Long liveId, Long categoryId, String title, String description, Integer price, Integer viewCount, Integer isSold, Long isLive) {
+    public Product(String userId, Integer liveId, Integer categoryId, String title, String description, Integer price, Integer viewCount, Integer isSold, Integer isLive) {
         this.userId = userId;
         this.liveId = liveId;
         this.categoryId = categoryId;
