+++ backend/src/main/java/com/ssafy/db/entity/Wish.java
@@ -22,10 +22,10 @@ import javax.persistence.Table;
 public class Wish {
     @Id
     @Column(name="wish_product_pk")
-    Long wishproductpk;
+    Integer wishproductpk;
 
     @Column(name="product_pk")
-    Long productpk;
+    Integer productpk;
 
     @Column(name="user_id_buyer")
     String useridbuyer;
